% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_DE_logp0.R
\name{DAISIE_DE_logp0}
\alias{DAISIE_DE_logp0}
\title{testing fuction, for comparison with DAISIE}
\usage{
DAISIE_DE_logp0(
  datalist,
  parameter,
  atol = 1e-15,
  rtol = 1e-15,
  methode = "ode45",
  rcpp_methode = "odeint::bulirsch_stoer",
  use_Rcpp = 0
)
}
\description{
This function compute the likelihood that all species that colonize the island
have gone extinct prior to the present.
}
\examples{
#Load DAISIE package and data
library(DAISIE)
data("Galapagos_datalist")
datalist <- Galapagos_datalist

parameter <- c(2.546591, 2.678781, Inf, 0.009326754, 1.008583)
# Compute log‐likelihood under the DE‐trait model
DAISIE_DE_logp0(
  datalist            = datalist,
  parameter           = parameter,
  atol                = 1e-15,
  rtol                = 1e-15,
  methode             = "ode45"
)
}
