% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_DE_logpES.R
\name{DAISIE_DE_logpES}
\alias{DAISIE_DE_logpES}
\title{This function calculates the likelihood of observing a non-endemic lineage with specified species trait states,}
\usage{
DAISIE_DE_logpES(
  brts,
  status,
  parameter,
  missnumspec,
  atol = 1e-15,
  rtol = 1e-15,
  methode = "ode45",
  rcpp_methode = "odeint::bulirsch_stoer",
  use_Rcpp = 0
)
}
\description{
This function calculates the likelihood of observing a non-endemic lineage with specified species trait states,
}
\examples{
library(DAISIE)
data("Galapagos_datalist")
datalist <- Galapagos_datalist

i <- 6
brts <- datalist[[i]]$branching_times

parameter <- c(2.546591, 2.678781, Inf, 0.009326754, 1.008583)
brts <- datalist[[i]]$branching_times
missnumspec <- datalist[[i]]$missing_species
DAISIE_DE_logpES(
  brts                    = brts,
  status                  = 2,
  parameter               = parameter,
  atol                    = 1e-15,
  missnumspec             = missnumspec,
  rtol                    = 1e-15,
  methode                 = "ode45"
)
}
