% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_DE_logpNE.R
\name{DAISIE_DE_logpNE}
\alias{DAISIE_DE_logpNE}
\title{This function calculates the likelihood of observing a non-endemic lineage with specified species trait states,}
\usage{
DAISIE_DE_logpNE(
  brts,
  status,
  parameter,
  sampling_fraction = NA,
  atol = 1e-15,
  rtol = 1e-15,
  methode = "ode45",
  rcpp_methode = "odeint::bulirsch_stoer",
  use_Rcpp = 0
)
}
\description{
This function calculates the likelihood of observing a non-endemic lineage with specified species trait states,
}
\examples{
library(DAISIE)
data("Galapagos_datalist")
datalist <- Galapagos_datalist

i <- 3
brts <- datalist[[i]]$branching_times

parameter <- c(2.546591, 2.678781, Inf, 0.009326754, 1.008583)
brts <- datalist[[i]]$branching_times
DAISIE_DE_logpNE(
  brts                    = brts,
  status                  = 4,
  parameter               = parameter,
  atol                    = 1e-15,
  rtol                    = 1e-15,
  methode                 = "ode45"
)
}
