% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_branch.R
\name{solve_branch}
\alias{solve_branch}
\title{branch solving}
\usage{
solve_branch(
  interval_func,
  initial_conditions,
  time,
  parameter,
  methode = "ode45",
  rcpp_methode = "odeint::bulirsch_stoer",
  atol,
  rtol,
  use_Rcpp = TRUE
)
}
\arguments{
\item{interval_func}{chosen function for interval, can also be string if using Rcpp}

\item{initial_conditions}{vector of initial conditions}

\item{time}{vector with two time points}
}
\description{
solve along branch
}
